+++ extensions/xcode/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Xcode Changelog
 
+
+## [Added Delete App Data] - 2024-09-13
+
+- Added delete app data command to remove the data of an app installed on a simulator without uninstalling.
+
 ## [Clear SwiftUI Previews Cache] - 2024-04-03
 
 - Added "Clear SwiftUI Previews Cache" command to remove the ~/Library/Developer/Xcode/UserData/Previews directory